feat: infra.go errors on err assignment
